docs(test): add benchmarking section and reorganize benchmark tests (#122)
* docs(test): add benchmarking section and reorganize benchmark tests

- Added a new **Benchmarking** section in `TESTING.md` to guide developers on performance testing.
- Moved `disk_test.go` into `test/benchmark/` and renamed its package to `benchmark` for clearer separation of benchmarking tests.


* docs(test): enhance test descriptions


* docs(test): update benchmarking section to include profiling instructions


---------
